move selectQuestions.js into utils dir
